+++ R/general.R
@@ -193,7 +193,7 @@ create_post_jointVIP <- function(object,
 #'
 #' @param object a jointVIP object
 #' @param ... not used
-#' @param smd specify the standardized mean difference is `OVB-based` or `standard`
+#' @param smd specify the standardized mean difference is `cross-sample` or `pooled`
 #' @param use_abs TRUE (default) for absolute measures
 #' @param bias_tol numeric 0.01 (default) any bias above the absolute bias_tol will be summarized
 #' @return no return value
@@ -222,7 +222,7 @@ create_post_jointVIP <- function(object,
 #' summary(new_jointVIP)
 summary.jointVIP <- function(object,
                              ...,
-                             smd = 'OVB-based',
+                             smd = 'cross-sample',
                              use_abs = TRUE,
                              bias_tol = 0.01) {
   if (any(is.null(names(list(...)))) & length(list(...)) > 0) {
@@ -268,7 +268,7 @@ summary.jointVIP <- function(object,
 #'
 #' @param object a post_jointVIP object
 #' @param ... not used
-#' @param smd specify the standardized mean difference is `OVB-based` or `standard`
+#' @param smd specify the standardized mean difference is `cross-sample` or `pooled`
 #' @param use_abs TRUE (default) for absolute measures
 #' @param bias_tol numeric 0.01 (default) any bias above the absolute bias_tol will be summarized
 #' @param post_bias_tol numeric 0.005 (default) any bias above the absolute bias_tol will be summarized
@@ -308,7 +308,7 @@ summary.jointVIP <- function(object,
 #' summary(post_dat_jointVIP)
 summary.post_jointVIP <- function(object,
                                   ...,
-                                  smd = 'OVB-based',
+                                  smd = 'cross-sample',
                                   use_abs = TRUE,
                                   bias_tol = 0.01,
                                   post_bias_tol = 0.005) {
@@ -341,7 +341,7 @@ summary.post_jointVIP <- function(object,
 #'
 #' @param x a jointVIP object
 #' @param ... not used
-#' @param smd specify the standardized mean difference is `OVB-based` or `standard`
+#' @param smd specify the standardized mean difference is `cross-sample` or `pooled`
 #' @param use_abs TRUE (default) for absolute measures
 #' @param bias_tol numeric 0.01 (default) any bias above the absolute bias_tol will be printed
 #'
@@ -371,7 +371,7 @@ summary.post_jointVIP <- function(object,
 #' print(new_jointVIP)
 print.jointVIP <- function(x,
                            ...,
-                           smd = 'OVB-based',
+                           smd = 'cross-sample',
                            use_abs = TRUE,
                            bias_tol = 0.01) {
   if (any(is.null(names(list(...)))) & length(list(...)) > 0) {
@@ -396,7 +396,7 @@ print.jointVIP <- function(x,
 #'
 #' @param x a post_jointVIP object
 #' @param ... not used
-#' @param smd specify the standardized mean difference is `OVB-based` or `standard`
+#' @param smd specify the standardized mean difference is `cross-sample` or `pooled`
 #' @param use_abs TRUE (default) for absolute measures
 #' @param bias_tol numeric 0.01 (default) any bias above the absolute bias_tol will be printed
 #'
@@ -436,7 +436,7 @@ print.jointVIP <- function(x,
 #' print(post_dat_jointVIP)
 print.post_jointVIP <- function(x,
                                 ...,
-                                smd = 'OVB-based',
+                                smd = 'cross-sample',
                                 use_abs = TRUE,
                                 bias_tol = 0.01) {
   if (use_abs) {
@@ -458,8 +458,8 @@ print.post_jointVIP <- function(x,
 #'
 #'
 #' @param x a jointVIP object
-#' @param ... custom options: `bias_curve_cutoffs`, `text_size`, `max.overlaps`, `label_cut_std_md`, `label_cut_outcome_cor`, `label_cut_bias`, `OVB_curves`, `add_var_labs`, `expanded_y_curvelab`
-#' @param smd specify the standardized mean difference is `OVB-based` or `standard`
+#' @param ... custom options: `bias_curve_cutoffs`, `text_size`, `max.overlaps`, `label_cut_std_md`, `label_cut_outcome_cor`, `label_cut_bias`, `bias_curves`, `add_var_labs`, `expanded_y_curvelab`
+#' @param smd specify the standardized mean difference is `cross-sample` or `pooled`
 #' @param use_abs TRUE (default) for absolute measures
 #' @param plot_title optional string for plot title
 #'
@@ -490,7 +490,7 @@ print.post_jointVIP <- function(x,
 #' plot(new_jointVIP)
 plot.jointVIP <- function(x,
                           ...,
-                          smd = 'OVB-based',
+                          smd = 'cross-sample',
                           use_abs = TRUE,
                           plot_title = "Joint Variable Importance Plot") {
   if (any(is.null(names(list(...)))) & length(list(...)) > 0) {
@@ -516,7 +516,7 @@ plot.jointVIP <- function(x,
           'label_cut_std_md',
           'label_cut_outcome_cor',
           'label_cut_bias',
-          'OVB_curves',
+          'bias_curves',
           'add_var_labs',
           'expanded_y_curvelab'
         )
@@ -530,7 +530,7 @@ plot.jointVIP <- function(x,
             ' label_cut_std_md',
             ' label_cut_outcome_cor',
             ' label_cut_bias',
-            ' OVB_curves',
+            ' bias_curves',
             ' add_var_labs',
             ' expanded_y_curvelab'
           )
@@ -545,7 +545,7 @@ plot.jointVIP <- function(x,
     measures = get_measures(x, smd = smd)
   }
   
-  if (smd == 'standard') {
+  if (smd == "pooled") {
     p <- ggplot(measures,
                 aes(x = .data$std_md,
                     y = .data$outcome_cor))  +
@@ -619,15 +619,15 @@ plot.jointVIP <- function(x,
       ), 2)))
   }
   
-  OVB_curves = list(...)[['OVB_curves']]
-  if (is.null(OVB_curves)) {
-    OVB_curves = TRUE
-  } else if (!is.logical(OVB_curves)) {
-    stop("`OVB_curves` can only be set as TRUE or FALSE")
+  bias_curves = list(...)[['bias_curves']]
+  if (is.null(bias_curves)) {
+    bias_curves = TRUE
+  } else if (!is.logical(bias_curves)) {
+    stop("`bias_curves` can only be set as TRUE or FALSE")
   }
-  if (OVB_curves) {
-    if (smd == "OVB-based") {
-      p <- add_OVB_curves(p,
+  if (bias_curves) {
+    if (smd == "cross-sample") {
+      p <- add_bias_curves(p,
                           use_abs = use_abs,
                           measures = measures, ...)
     }
@@ -652,7 +652,7 @@ plot.jointVIP <- function(x,
 #' @param ... encompasses other variables needed
 #' @return a joint variable importance plot of class `ggplot` with curves
 #' @import ggplot2
-add_OVB_curves <- function(p, ...) {
+add_bias_curves <- function(p, ...) {
   use_abs = list(...)[['use_abs']]
   measures = list(...)[['measures']]
   bias_curve_cutoffs = list(...)[['bias_curve_cutoffs']]
@@ -903,8 +903,8 @@ floor_dec <- function(num, dec_place = 1) {
 #'
 #'
 #' @param x a post_jointVIP object
-#' @param ... custom options: `bias_curve_cutoffs`, `text_size`, `max.overlaps`, `label_cut_std_md`, `label_cut_outcome_cor`, `label_cut_bias`, `OVB_curves`, `add_var_labs`, `expanded_y_curvelab`
-#' @param smd specify the standardized mean difference is `OVB-based` or `standard`
+#' @param ... custom options: `bias_curve_cutoffs`, `text_size`, `max.overlaps`, `label_cut_std_md`, `label_cut_outcome_cor`, `label_cut_bias`, `bias_curves`, `add_var_labs`, `expanded_y_curvelab`
+#' @param smd specify the standardized mean difference is `cross-sample` or `pooled`
 #' @param use_abs TRUE (default) for absolute measures
 #' @param plot_title optional string for plot title
 #' @param add_post_labs TRUE (default) show post-measure labels
@@ -948,7 +948,7 @@ floor_dec <- function(num, dec_place = 1) {
 #' plot(post_dat_jointVIP)
 plot.post_jointVIP <- function(x,
                                ...,
-                               smd = 'OVB-based',
+                               smd = 'cross-sample',
                                use_abs = TRUE,
                                plot_title = "Joint Variable Importance Plot",
                                add_post_labs = TRUE,
@@ -986,7 +986,7 @@ plot.post_jointVIP <- function(x,
     p$layers[[length(p$layers)]] <- NULL
   }
   
-  if (smd == 'standard') {
+  if (smd == "pooled") {
     p <- p + geom_point(data = post_measures,
                         aes(x = .data$post_std_md,
                             y = .data$outcome_cor))
@@ -1024,7 +1024,7 @@ plot.post_jointVIP <- function(x,
       post_measures[!(abs(round(post_measures$post_bias, 4)) >= post_label_cut_bias), 'text_label'] = ""
     }
     
-    if (smd == 'standard') {
+    if (smd == "pooled") {
       p <- p + geom_text_repel(
         data = post_measures,
         aes(
@@ -1072,8 +1072,8 @@ plot.post_jointVIP <- function(x,
 #'
 #'
 #' @param x a jointVIP object
-#' @param ... custom options: `bias_curve_cutoffs`, `text_size`, `max.overlaps`, `label_cut_std_md`, `label_cut_outcome_cor`, `label_cut_bias`, `OVB_curves`, `add_var_labs`
-#' @param smd specify the standardized mean difference is `OVB-based` or `standard`
+#' @param ... custom options: `bias_curve_cutoffs`, `text_size`, `max.overlaps`, `label_cut_std_md`, `label_cut_outcome_cor`, `label_cut_bias`, `bias_curves`, `add_var_labs`
+#' @param smd specify the standardized mean difference is `cross-sample` or `pooled`
 #' @param use_abs TRUE (default) for absolute measures
 #' @param plot_title optional string for plot title
 #' @param B 100 (default) for the number of times the bootstrap step wished to run
@@ -1108,7 +1108,7 @@ plot.post_jointVIP <- function(x,
 #' bootstrap.plot(new_jointVIP, B = 15)
 bootstrap.plot <- function(x,
                            ...,
-                           smd = 'OVB-based',
+                           smd = 'cross-sample',
                            use_abs = TRUE,
                            plot_title = "Joint Variable Importance Plot",
                            B = 100) {
@@ -1116,17 +1116,17 @@ bootstrap.plot <- function(x,
     stop("bootstrap_plot function only applicable to class jointVIP only!")
   }
   
-  OVB_curves = list(...)[['OVB_curves']]
-  if (is.null(OVB_curves)) {
-    if (smd == 'OVB-based') {
-      specified_OVB_curves = TRUE
+  bias_curves = list(...)[['bias_curves']]
+  if (is.null(bias_curves)) {
+    if (smd == "cross-sample") {
+      specified_bias_curves = TRUE
     } else {
-      specified_OVB_curves = FALSE
+      specified_bias_curves = FALSE
     }
-  } else if (!is.logical(OVB_curves)) {
-    stop("`OVB_curves` can only be set as TRUE or FALSE")
+  } else if (!is.logical(bias_curves)) {
+    stop("`bias_curves` can only be set as TRUE or FALSE")
   } else {
-    specified_OVB_curves = OVB_curves
+    specified_bias_curves = bias_curves
   }
   
   p <-
@@ -1134,7 +1134,7 @@ bootstrap.plot <- function(x,
       x,
       ...,
       smd = smd,
-      OVB_curves = FALSE,
+      bias_curves = FALSE,
       use_abs = use_abs,
       plot_title = plot_title
     )
@@ -1186,8 +1186,8 @@ bootstrap.plot <- function(x,
       alpha = 0.4
     )
   
-  if (smd == "OVB-based" & specified_OVB_curves) {
-    p <- add_OVB_curves(
+  if (smd == "cross-sample" & specified_bias_curves) {
+    p <- add_bias_curves(
       p,
       use_abs = use_abs,
       measures = og,